Moved roslib module extension to its own file
