+++ src/declarations/types.d.ts
@@ -1,0 +1,1 @@
+export type VoidFunc = () => void;
